+++ cub_src/parse/set_map.c
@@ -6,7 +6,7 @@
 /*   By: mforstho <mforstho@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/02/27 14:08:51 by mforstho      #+#    #+#                 */
-/*   Updated: 2023/03/03 19:40:55 by ageels        ########   odam.nl         */
+/*   Updated: 2023/03/06 20:12:15 by ageels        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,15 +18,15 @@ void	set_tile(t_data *data, t_par *par, int i, int *j)
 
 	c = par->maparray[i][*j];
 	if (c == ' ' || c == 'N' || c == 'S' || c == 'E' || c == 'W')
-		data->map[*j + (data->max.x * i)] = 0;
+		data->map[(int)(*j + (data->max.x * i))] = 0;
 	else
-		data->map[*j + (data->max.x * i)] = par->maparray[i][*j] - '0';
+		data->map[(int)(*j + (data->max.x * i))] = par->maparray[i][*j] - '0';
 	if (par->maparray[i][*j + 1] == '\0')
 	{
 		(*j)++;
 		while (*j < data->max.x)
 		{
-			data->map[(*j + (data->max.x * i))] = 0;
+			data->map[(int)(*j + (data->max.x * i))] = 0;
 			(*j)++;
 		}
 	}
